move Loader class into spark_preprocess module
